Moved BiFunction into the 'threeten' package since it's only used there.
